Improved Traffic Analytics UI
ref https://linear.app/ghost/issue/PROD-1639/analytics-ui-refinements

- TBD
